requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
colorama==0.4.6
pygit2==1.13.3
jwt==1.3.1
PyGithub==2.1.1
python-dotenv==1.0.1
python-dateutil==2.8.2
supabase==2.3.5
sendgrid==6.11.0
slack_bolt==1.18.1